requirements.txt
aiohttp==1.0.5
async-timeout==1.1.0
cffi==1.8.3
chardet==2.3.0
discord.py==0.14.2
ez-setup==0.9
httplib2==0.9.2
multidict==2.1.2
oauth2client==4.0.0
opuslib==1.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyNaCl==1.0.1
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
websockets==3.2
youtube-dl==2016.11.4
aiohttp==1.0.5
async-timeout==1.1.0
cffi==1.8.3
chardet==2.3.0
decorator==4.0.10
discord.py==0.14.2
ez-setup==0.9
httplib2==0.9.2
multidict==2.1.2
oauth2client==4.0.0
opuslib==1.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyNaCl==1.0.1
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
websockets==3.2
youtube-dl==2016.11.4
